requirements.txt
google_search_results==2.4.2
python-dotenv==1.1.1